refactor(web): add password validation to form
